Add option to make Length unit optional
This seems under-specified in SVG2, but for the <svg> x/y/width/height
the unit was optional in SVG1.1, so width="100" is valid. In SVG2, it
references the CSS3 <length-percentage> dimension, which requires the
unit if the number is non-zero.

However, since the SVG2 spec has examples where it doesn't provide the
parameter, and in the interest of backcompat, make the unit optional
(but make the behavior configurable).