Add short-string optimization to RcString
This avoids allocations if the string is shorter than the storage it
would take to store the shared_ptr state, by storing the string inline.

This uses an approach similar to libc++, which maximizes the number of
characters that can be stored and only requires one byte to hold the
short string size.

Additionally, add an lldb command script to pretty-print RcString in the
debugger.