Upgrade toolchain to Clang 13
libc++13 has better support for ranges.

To enable the upgrade, downgrade the devcontainer to Ubuntu Bionic,
since the previous version was introducing dependency issues because it
uses libpython3.9, and the liblldb package for some reason has a
dependency on libpython3.8, which cannot be resolved.  Note that this is
two versions back, since this is the only other one that also supports
arm64 for M1 macs.

Additional changes:
* Upgrade gtest to v1.11.0, since the previous version was relying on
   a the deprecated `std::result_of` API, which is removed in C++20.
   This version of libc++ removes the API.
   See google/googletest#2914
* Switch to an embedded clang toolchain instead of requiring it to be
   pre-installed.
* Switch CI to bionic to match devcontainer.