Start using resvg-test-suite
This contains an extensive set of tests for SVG1.1, specifically
for static-only, which is exactly what we want to test.

Create a framework to allow comparing these images, breaking them
into parameter-driven test suites, so that we can test each feature
in groups or with individual tests.

Also fix a few bugs necessary to run the tests:
- Support 'currentColor' to fix a crash about bad variant access
- Change Skia to render to an unpremultiplied target, since stb expects
  that.
- Support overriding the output size to 500x500, to match golden images.
  Why this needs to be done is still to be debugged.

Other test failures have been categorized as either bugs or known
missing functionality in `resvg_test_suite.cc`.