Add support for gradient href attributes
On <linearGradient> and <radialGradient>, if there's an href attribute
it enables inheriting attributes or <stop> elements from other elements.

For <stop> elements, these are inherited only if the current element has
only descriptive elements within, like <desc>, <metadata>, <title>.

For attributes, these may be inherited independent of <stop> element
conditions, either from a matching gradient type or different type.

To do this, instantiate a shadow tree if <stop> elements are inherited,
but then manually inherit attributes through the chain.  To handle cases
of recursion, add a guard to prevent referencing the same elememnt in
the chain.